ml-service/requirements.txt
# TensorFlow CPU — image Docker ~1.5 Go ; prévoir plan Render avec assez de RAM (ex. 2 Go).
tensorflow-cpu>=2.16.0,<2.20
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
